pyproject.toml
[project]
name = "cefp-kms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "cryptography>=44.0.3",
    "datetime>=5.5",
    "flasgger>=0.9.7.1",
    "flask>=3.1.1",
    "flask-cors>=6.0.0",
    "flask-limiter>=3.11.0",
    "flask-sqlalchemy>=3.1.1",
]
